support variable k in one batch

+++ csrc/sampling.cu
@@ -309,7 +309,7 @@ void radik_sampling_from_probs(at::Tensor workspace_buffer, at::Tensor probs, at
   cudaError_t status = sampling::RadiKSamplingFromProb<float, int>(
       static_cast<float*>(probs.data_ptr()), static_cast<int*>(output.data_ptr()),
       maybe_indices.has_value() ? static_cast<int*>(maybe_indices->data_ptr()) : nullptr,
-      has_top_k_arr ? static_cast<float*>(maybe_top_k_arr->data_ptr()) : nullptr, batch_size,
+      has_top_k_arr ? static_cast<int*>(maybe_top_k_arr->data_ptr()) : nullptr, batch_size,
       top_k_val, vocab_size, philox_seed, philox_offset, workspace_buffer.data_ptr(),
       workspace_buffer.element_size() * workspace_buffer.size(0), stream);
   TORCH_CHECK(status == cudaSuccess, "RadiKSamplingFromProbs failed with error code " +
